+++ packages/design-system/src/lib/stories/select-advanced/SelectAdvanced.stories.tsx
@@ -92,3 +92,16 @@ export const WithDifferentSelectWord: Story = {
     </CanvasFixedHeight>
   )
 }
+
+export const WithoutPlaceholderOptionInMenu: Story = {
+  render: () => (
+    <CanvasFixedHeight height={250}>
+      <SelectAdvanced
+        options={exampleOptions}
+        locales={{ select: 'Select something...' }}
+        showPlaceholderOptionInMenu={false}
+        isSearchable={false}
+      />
+    </CanvasFixedHeight>
+  )
+}